[FAB-1652] Split docker image up into multiple tiers
We need to address the static-binary/dns problem introduced by commit
b61723f9.  We also want to look forward to the point where we want
to optimize JVM based payloads, such as java chaincode.  Therefore
we split the baseimage up into three tiers:

*) baseos - a basic, albeit $arch normalized base os
*) basejvm - derived from baseos + adds a JVM environment
*) baseimage - derived from basejvm, adds all the other stuff we've always had.

From the perspective of baseimage consumers, nothing has changed.  They
consume baseimage as they always have and it looks/behaves identitcally.

The difference is, we now have the option to reference either baseos or basejvm.
The short-term plan is that we will replace the current runtime/busybox based
mechanism employed in fabric/fabric-cop with baseos.  This will have an
impact in two ways.

1) The size of the image will increase from ~5MB to ~155MB (uncompressed).
2) DNS will once again function correctly.

While (1) is unfortunate, its also not horrible.  In addition, we can possibly
further optimize this in the future by basing on something like alpine rather
than Ubuntu, TBD.  We also gain the flexibility to support JVM payloads in
the future.  This would have been difficult under the current busybox
trajectory.

Change-Id: I5e3475ffed3be905184a5f82ca94e828c42ca9b6
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>