[FABB-151] Remove delayed_commits=true
Since ensure_full_commit API in CouchDB doesn't work as documented in CouchDB v2.x,
Fabric requires delayed_commits to be false.
'false' is the default in CouchDB, therefore this commit simply removes the
override of delayed_commits=true.
Using the default value of 'false' ensures that writes are synced to disk.

Signed-off-by: David Enyeart <enyeart@us.ibm.com>